Lab 7/requirements.txt
fastapi
uvicorn
pydantic
typing
mysql-connector-python